Document transition from experimental to unstable
Upstream is now no longer explicitly naming its releases with the
"experimental-" prefix.

The Wireguard project is still not ready to commit to long-term
support for the wire format, or backporting security problems
discovered in the future, so we don't want to unleash wireguard on
debian stable.

But it's demonstrated enough reliability that I think it can be
distributed in debian unstable.  Moving it to unstable also gives it
wider visibility, which hopefully can help upstream discover and fix
any other problems.

These changes prepare for the transition from debian experminental to
debian unstable.