loosen versioned dependencies of wireguard metapackage
wireguard-dkms is Arch: all, while wireguard-tools is Arch: any.

This means that for rolling-release distributions that support
multiple architectures, sometimes a machine with an architectures with
the slower build daemon will have the Arch: all packages available
before the Arch: any packages.

For the wireguard metapackage, the tight version dependencies meant
that in a rolling release, users who tried to do "apt install
wireguard" would sometimes find themselves in an uninstallable state.

The wireguard metapackage is still tightly-coupled to the version of
the wireguard-dkms package, since they're both Arch: all, and i didn't
want someone to be able to have wireguard 0.0.20180606-1 installed but
wireguard-dkms 0.0.20180202-1.  But i've loosened the dependency on
wireguard-tools to the first version that supports the netlink
interface.

Note that since systemd 237 supports wireguard, even the
wireguard-tools dependency is strictly unnecessary, but for folks who
want the most minimal system, they can just avoid the metapackage
entirely.

This decision arose out of on-list discussion between several people:

https://lists.zx2c4.com/pipermail/wireguard/2018-January/002321.html