Import module reload postinst script from wireguard package
The wireguard metapackage used to host the postinst script which would
reload the kernel module on upgrade.

But since the package split happened, and the wireguard metapackage
comes from the other source, there is no guarantee that the cadence of
the upgrade of the metapackage will coincide with the cadence of the
kernel module package.

So we move the reload script into wireguard-dkms.