Avoid asking for reboot unnecessarily if old module is not loaded
We should not signal that a reboot is required if the old module is
not loaded, even if the wireguard package is not installed.

There is still a case that we could probably handle safely, but we
don't:

 - wireguard package not installed
 - old module loaded
 - "wg show interfaces" returns the empty string

We should be able to remove the old module and load the new one safely
in this case, but doing so complicates the story about the semantics
of having the "wireguard" package installed, to the point where people
are likely to be confused.  So we keep it simple, and just mark reboot
required instead in that case.

Signed-off-by: Daniel Kahn Gillmor <dkg@fifthhorseman.net>