postinst: remove bespoke error-prone auto-update mechanism
As discussed at length, relying on a metapackage to control update
behavior is confusing and unclear, and given the popularity of the
metapackage, this essentially opts users in to this broken behavior.

This is not behavior many people really want. There might be all sorts
of weird configurations that simply do not react well to a forced
removal of the kernel module, and unlike actual userspace daemons like
sshd, which fork off a process for each login, bringing down the
wireguard interface means fully severing the connection. This means that
people ssh'd in and performing updates will be disconnected for up to 15
seconds (if they even manage to reconnect), depending on the
configuration of the server. It's simply not sensible to force this upon
users.

While this kind of mechanism might have made sense (and as an opt-in
only, never an opt-out like this one) in the early days of wireguard,
when releases were frequent and also occasionally buggy, at this late
stage we should probably start taking things a bit more seriously.

Not only is this behavior the wrong thing for many users, but it is also
completely non-standard for kernel modules in general. There is no other
kernel module in Debian that has this kind of odd behavior. The standard
way of dealing with this is via the reboot-required files, which is what
this commit changes the postinst script to use.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>